Move and rename Guzzle\AuthenticatorSubscriber to HttpClient\Authenticator
